# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart School/Madrasa ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted

  # ── CLASSES ───────────────────────────────────────────────────────────────
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/StudentStatus"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
    post:
      operationId: createStudent
      tags: [students]
      summary: Admit a new student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student admitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Student deleted

  # ── ATTENDANCE ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      enum: [SUPER_ADMIN, TEACHER, ACCOUNTANT, PARENT, STUDENT]
    StudentStatus:
      type: string
      enum: [ACTIVE, INACTIVE, GRADUATED, TRANSFERRED]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required:
        - users
        - total
...
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required:
        - email
        - password
        - firstName
        - lastName
        - role
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        isActive:
          type: boolean

    # ── CLASS ─────────────────────────────────────────────────────────────
...
    Student:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
          nullable: true
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          nullable: true
        address:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        parentName:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        parentEmail:
          type: string
          format: email
          nullable: true
        classId:
          type: integer
          nullable: true
        className:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/StudentStatus"
        admissionDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - firstName
        - lastName
        - status
        - admissionDate
        - createdAt
    StudentListResponse:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
      required:
        - students
        - total
    CreateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        address:
          type: string
        phoneNumber:
          type: string
        parentName:
          type: string
        parentPhone:
          type: string
        parentEmail:
          type: string
          format: email
        classId:
          type: integer
        admissionDate:
          type: string
          format: date
      required:
        - firstName
        - lastName
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        address:
          type: string
        phoneNumber:
          type: string
        parentName:
          type: string
        parentPhone:
          type: string
        parentEmail:
          type: string
          format: email
        classId:
          type: integer
        status:
          $ref: "#/components/schemas/StudentStatus"

    # ── ATTENDANCE ────────────────────────────────────────────────────────
...